+++ src/lang/dsl/mod.rs
@@ -11,6 +11,7 @@
 // pub mod j;
 // #[cfg(feature = "alloc")]
 // pub mod php;
+// pub mod _wip_lisp;
 
 crate::structural_mods! { // _mods
     _mods {
@@ -21,5 +22,6 @@ crate::structural_mods! { // _mods
         // pub use super::j::_all::*;
         // #[cfg(feature = "alloc")]
         // pub use super::php::_all::*;
+        // pub use super::_wip_lisp::*;
     }
 }