+++ src/lib.rs
@@ -216,9 +216,11 @@ mod _hidden {
 pub(crate) use _crate_internals::*;
 mod _crate_internals {
     pub(crate) use devela_base_core::_workspace_internals::*;
+    // #[cfg(feature = "devela_base_data")]
     // pub(crate) use devela_base_data::_workspace_internals::*;
     pub(crate) use devela_base_num::_workspace_internals::*;
-    // pub(crate) use devela_base_text::_workspace_internals::*;
+    #[cfg(feature = "devela_base_text")]
+    pub(crate) use devela_base_text::_workspace_internals::*;
 
     #[allow(unused_imports)]
     #[rustfmt::skip]